fix: change userId type from String to Long to match UserDO

+++ weindependent-app/src/main/java/com/weindependent/app/database/mapper/weindependent/UserMapper.java
@@ -14,7 +14,8 @@ public interface UserMapper {
     UserDO findByAccount(String account);
     int insert(UserDO user);
     UserDO findById(String id);
-    int updatePassword(@Param("id") int id, @Param("password") String password);
+    // int updatePassword(@Param("id") int id, @Param("password") String password);
+    int updatePassword(@Param("id") Long id, @Param("password") String password);
     boolean existsById(int id);
     UserDO findById(Long userId); //Hurely add for pdf download
     void updateUser(UserDO userDO);
